Added OpModes and Supporting Classes
NewArmControl is the main opMode and coordinator of subroutines.
OneSidedController does most of the UI work, is used by NewArmControl and uses Robot_ and NewArmMath3.
NewArmMath3 does all of the math, calibration, and integration over time of input for the arms, but is completely isolated from the FTC app (i.e. all calls made too and from this program refer only to code I wrote).
Robot_ is a proxy for HardwareMap that is build specifically for out use (i.e. it would have been reasonable to make Robot_ extend HardwareMap) It handles initialization.
StallDetector is called directly from NewArmControl, it has no effect on the remainder of the program whatsoever with the exception of throwing a helpful error message and terminating the program whenever a moter is at risk of burning out.
SelfTest is a simple OpMode that only uses Robot_ and is intended for checking if components are plugged in correctly and isolating bugs that lie outside the scope of my code.